fix travis tests, CreateExportSpec out of AppSpec
